pyproject.toml
[project]
name = "claude-util"
version = "0.1.0"
description = "CLI tool for managing multiple Claude Code subscription accounts"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0,<14",
  "questionary>=2.0.1,<3",
  "pydantic>=2.6.0,<3",
  "pyyaml>=6.0.1,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9",
  "pytest-cov>=4.1.0,<5",
  "mypy>=1.8.0,<2",
  "ruff>=0.2.0",
  "pre-commit>=3.6.0,<4",
  "types-pyyaml>=6.0.12,<7",
]

[project.scripts]
claude-cli = "claude_cli.main:app"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"claude_cli" = "src"}
packages = [
  "claude_cli", "claude_cli.commands", "claude_cli.core",
  "claude_cli.models", "claude_cli.ui", "claude_cli.utils"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
